% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_time.R
\name{big_time_op}
\alias{big_time_op}
\title{Big Time Function Operator}
\usage{
big_time_op(f)
}
\arguments{
\item{f}{Function to wrap}
}
\value{

}
\description{
Wrap a function of your choosing with the "this shit is big time" YMH drop
}
\examples{
\dontrun{
x <- 1:10

big_time_mean <- big_time_op(mean)

big_time_mean(x)
}
}
