% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ee_post.R
\name{create_ee_post}
\alias{create_ee_post}
\title{Create a blog post with ee specs}
\usage{
create_ee_post(..., open = TRUE)
}
\arguments{
\item{...}{arguments passed to \code{distill::create_post()}}

\item{open}{logical. Open the created file?}
}
\value{

}
\description{
This function wraps \code{distill::create_post()} to set options/load libraries in the resulting file
}
\examples{
\dontrun{
library(eemisc)
create_ee_post("Test Post")
}
}
