% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{\%!in\%}
\alias{\%!in\%}
\title{%!in% (not in)}
\usage{
lhs \%!in\% rhs
}
\arguments{
\item{lhs}{left-hand side}

\item{rhs}{right-hand size}
}
\value{

}
\description{
A more intuitive way to test for x not in y
}
\examples{
\dontrun{
a <- c("a", "b", "c", "d")
b <- c("a", "b")
a \%!in\% b
}
}
