% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{make_perc}
\alias{make_perc}
\title{Create a percentage as character}
\usage{
make_perc(num, denom, digits = 2)
}
\arguments{
\item{num}{numerator}

\item{denom}{denominator}

\item{digits}{decimal points to return}
}
\value{

}
\description{
Create a percentage as character
}
\examples{
\dontrun {
library(magrittr)
library(dplyr)

mtcars \%>\%
  mutate(perc = make_perc(cyl, mpg))
}
}
