% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_table}
\alias{make_table}
\title{Wrapper to quickly make html tables for blog}
\usage{
make_table(x)
}
\arguments{
\item{x}{a dataframe or tibble}
}
\value{

}
\description{
Wrapper to quickly make html tables for blog
}
\examples{
\dontrun{
library(knitr)
library(kableExtra)
library(magrittr)

#table
mtcars \%>\%
   select(mpg:disp) \%>\%
   make_table()
}
}
