% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{replace_null}
\alias{replace_null}
\title{Replace NULL}
\usage{
replace_null(x, replacement = NA_character_)
}
\arguments{
\item{x}{element to (potentially) replace}

\item{replacement}{replacement for NULL. Default is NA_character_}
}
\value{

}
\description{
This function provides a simple way to replace NULLs. It's will be most useful in conjunction with {purrr} functions when working with lists
}
\examples{
\dontrun{
a <- NULL
replace_null(a)
}
}
