% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-eedark.R
\name{theme_eedark}
\alias{theme_eedark}
\title{[ggplot2] theme based loosely on theme_ipsum from hrbrthemes.}
\usage{
theme_eedark(
  size = 10,
  base_font = "sans",
  base_color = "#cfd8dc",
  accent_color = "#717786",
  faceted = FALSE
)
}
\arguments{
\item{size}{base font size}

\item{base_font}{base font family}

\item{base_color}{color of text and ticks}

\item{accent_color}{color of grid lines}

\item{faceted}{logical. TRUE will reduce margins to make facets larger}
}
\value{

}
\description{
[ggplot2] theme based loosely on theme_ipsum from hrbrthemes.
}
\examples{
\dontrun{
library(ggplot2)
library(dplyr)

#basic scatterplot
ggplot(mtcars, aes(mpg, wt)) +
  geom_point() +
  labs(x = "miles per gallon",
       y = "weight",
       title = "Scatterplot example",
       subtitle = "this is a subtitle",
       caption = "by bret bretterson") +
 theme_eedark()
}

}
\seealso{
\code{link{theme_ee}}
}
